Improved instructions to include namespace
This document does not explicitly called out the namespace that is required to be linked with the IAM role having the policy **CloudWatchAgentServerPolicy** attached. Because of that, the users may create such role and the linked service account in a different namespace because the link to the document given here is a generic link that asks the users to create the service account in default namespace. Later in the document, the instructions deploys the CloudWatch and Fluent Bit agents in a namespace called `amazon-cloudwatch`. This mismatch does not grant required permissions to the deployed pods to communicate with the corresponding CloudWatch service on the AWS account. 

This little change will callout this namespace requirements explicitly to avoid any confusion or pain later, which would be otherwise almost unavoidable. 

Contact parthpg@amazon.com for more details if required.